# Optimize `any::swap`