--- a/benchmarks/CMakeLists.txt
+++ b/benchmarks/CMakeLists.txt
@@ -102,6 +102,7 @@ endfunction()
 
 add_benchmark(adjacent_difference src/adjacent_difference.cpp)
 add_benchmark(adjacent_find src/adjacent_find.cpp)
+add_benchmark(any_swap src/any_swap.cpp)
 add_benchmark(bitset_from_string src/bitset_from_string.cpp)
 add_benchmark(bitset_to_string src/bitset_to_string.cpp)
 add_benchmark(efficient_nonlocking_print src/efficient_nonlocking_print.cpp)