--- a/stl/inc/any
+++ b/stl/inc/any
@@ -230,7 +230,9 @@ public:
     }
 
     void swap(any& _That) noexcept {
-        _That = _STD exchange(*this, _STD move(_That));
+        any _Old = _STD move(*this);
+        *this    = _STD move(_That);
+        _That    = _STD move(_Old);
     }
 
     // Observers [any.observers]