This pr removes an unnecessary copy in `any::swap`.

Hopefully fixes https://github.com/microsoft/STL/issues/5698.

Benchmark result:
```
------------------------------------------------------
Benchmark            Time             CPU   Iterations
------------------------------------------------------
bm<trivial>       14.5 ns         14.8 ns     49777778
bm<small>         21.1 ns         21.0 ns     32000000
bm<large>         15.2 ns         15.0 ns     44800000

->
------------------------------------------------------
Benchmark            Time             CPU   Iterations
------------------------------------------------------
bm<trivial>       10.1 ns         9.52 ns     64000000
bm<small>         15.9 ns         15.3 ns     40727273
bm<large>         10.6 ns         10.7 ns     64000000

```